+++ .gitignore
@@ -3,5 +3,6 @@
 node_modules
 package-lock.json
 *.lock
+*.log
 .vscode/
 coverage
