+++ api/index.js
@@ -1,15 +1,17 @@
 require("dotenv").config();
-const {
+import React from "preact";
+import renderToString from "preact-render-to-string";
+import {
   renderError,
   parseBoolean,
   parseArray,
   clampValue,
   CONSTANTS,
-} = require("../src/utils");
-const fetchStats = require("../src/fetchStats");
-const renderStatsCard = require("../src/renderStatsCard");
+} from "../src/utils";
+import fetchStats from "../src/fetch/stats";
+import statsCard from "../src/components/statsCard";
 
-module.exports = async (req, res) => {
+export default async (req, res) => {
   const {
     username,
     hide,
@@ -32,7 +34,7 @@ module.exports = async (req, res) => {
   try {
     stats = await fetchStats(username);
   } catch (err) {
-    return res.send(renderError(err.message));
+    return res.send(renderToString(renderError(err.message)));
   }
 
   const cacheSeconds = clampValue(
@@ -44,18 +46,20 @@ module.exports = async (req, res) => {
   res.setHeader("Cache-Control", `public, max-age=${cacheSeconds}`);
 
   res.send(
-    renderStatsCard(stats, {
-      hide: parseArray(hide),
-      show_icons: parseBoolean(show_icons),
-      hide_title: parseBoolean(hide_title),
-      hide_border: parseBoolean(hide_border),
-      hide_rank: parseBoolean(hide_rank),
-      line_height,
-      title_color,
-      icon_color,
-      text_color,
-      bg_color,
-      theme,
-    })
+    renderToString(
+      statsCard(stats, {
+        hide: parseArray(hide),
+        show_icons: parseBoolean(show_icons),
+        hide_title: parseBoolean(hide_title),
+        hide_border: parseBoolean(hide_border),
+        hide_rank: parseBoolean(hide_rank),
+        line_height,
+        title_color,
+        icon_color,
+        text_color,
+        bg_color,
+        theme,
+      })
+    )
   );
 };
