+++ src/utils/index.js
@@ -1,26 +1,44 @@
-const axios = require("axios");
-const themes = require("../themes");
+import React from "preact";
+import axios from "axios";
+import themes from "../../themes";
 
 const renderError = (message) => {
-  return `
-    <svg width="495" height="100" viewBox="0 0 495 100" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <style>
-    .text { font: 600 16px 'Segoe UI', Ubuntu, Sans-Serif; fill: #2F80ED }
-    .small { font: 600 12px 'Segoe UI', Ubuntu, Sans-Serif; fill: #252525 }
-    </style>
-    <rect x="0.5" y="0.5" width="494" height="99%" rx="4.5" fill="#FFFEFE" stroke="#E4E2E2"/>
-    <text x="25" y="45" class="text">Something went wrong! file an issue at https://git.io/JJmN9</text>
-    <text id="message" x="25" y="65" class="text small">${message}</text>
+  return (
+    <svg
+      width="495"
+      height="100"
+      viewBox="0 0 495 100"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <style>{`
+        .text {
+          font: 600 16px "Segoe UI", Ubuntu, Sans-Serif;
+          fill: #2f80ed;
+        }
+        .small {
+          font: 600 12px "Segoe UI", Ubuntu, Sans-Serif;
+          fill: #252525;
+        }
+      `}</style>
+      <rect
+        x="0.5"
+        y="0.5"
+        width="494"
+        height="99%"
+        rx="4.5"
+        fill="#FFFEFE"
+        stroke="#E4E2E2"
+      />
+      <text x="25" y="45" class="text">
+        Something went wrong! file an issue at https://git.io/JJmN9
+      </text>
+      <text id="message" x="25" y="65" class="text small">
+        {message}
+      </text>
     </svg>
-  `;
+  );
 };
-
-// https://stackoverflow.com/a/48073476/10629172
-function encodeHTML(str) {
-  return str.replace(/[\u00A0-\u9999<>&](?!#)/gim, function (i) {
-    return "&#" + i.charCodeAt(0) + ";";
-  });
-}
 
 function kFormatter(num) {
   return Math.abs(num) > 999
@@ -83,7 +101,7 @@ function FlexLayout({ items, gap, direction }) {
     if (direction === "column") {
       transform = `translate(0, ${gap * i})`;
     }
-    return `<g transform="${transform}">${item}</g>`;
+    return <g transform={transform}>{item}</g>;
   });
 }
 
@@ -132,10 +150,9 @@ const CONSTANTS = {
   ONE_DAY: 86400,
 };
 
-module.exports = {
+export {
   renderError,
   kFormatter,
-  encodeHTML,
   isValidHexColor,
   request,
   parseArray,
