+++ themes/index.js
@@ -79,4 +79,4 @@ const themes = {
   },
 };
 
-module.exports = themes;
+export default themes;
